buffer: make buffer::exception classes undefined in dynamic objects
On OSX, two classes with same name, but are privated to their own
objectss are not considred the same. This causes problem for exception
handling in ceph dynamic objects. The object class code try using
buffer::error to catch exceptions throwed by bufferlist::iterator.
But buffer::error in dynamic object and buffer::error in ceph-osd
are not considered the same. (bufferlist::iterator is defined in
ceph-osd).

The fix is force buffer::exception classes undefined in dynamic objects.
When dynamic objects are loaded into ceph-osd, buffer::exception classes
in ceph-used are used.

Signed-off-by: Yan, Zheng <zyan@redhat.com>