WorkQueue: correct the behaviour of ThreadPool::drain()
the existing ThreadPool::drain(wq) will wait until *all* the work
queues added to this pool are empty, which renders the `wq` parameter
useless. and the intended behavour of `ThreadPool::drain(wq)` should
be drain the wq *only*. this change does not "fix" the existing
behaviour of WorkQueue or improve the performance of its users.
instead, it addresses the correctness of `ThreadPool::drain(wq)`.

Fixes: #12444
Signed-off-by: Kefu Chai <kchai@redhat.com>